requirements.txt
fastapi
uvicorn
google-generativeai
pandas
scikit-learn
python-multipart
